Used hg convert to filter out the __NOTEBOOK__ directory and updated .hgignore.
The __NOTEBOOK__ directory primarily contained the binary index database and got
regenerated automatically after deletion. It also got updated continuously
showing as a change even when there where no changes in the actual notebook
contents.

Unfortunately KeepNote doesn't seem to be able to regenerate it under all
circumstances. In those cases one can create a new notebook and copy-paste the
contents of the old one in the file system.