Finished, transcribing, translating and organizing the draft definition sheets.
Some time ago a draft definition of a subset of Arcadia with some possible
extensions was created with the intention of making it small enough to be
implemented as part of the "Compilers" course of NTUA, yet powerful enough to
showcase a few key features.

Those A4 sheets together with the notes from discussing the subset with
professor Papaspyrou were transcribed and organized here.